CustomerEmbedding/requirements.txt
openai
pandas
numpy
SQLAlchemy
pymysql
python-dotenv
sqlmodel
cryptography
jinja2
python-multipart